Move splash related code to splash.c
